% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\name{dataviz}
\alias{dataviz}
\title{Data Viz}
\usage{
dataviz(dataset)
}
\arguments{
\item{dataset}{dataset for analysis}
}
\description{
Function that generates App in Shiny for analysis of input dataset
}
\examples{
if(interactive()) dataviz(mtcars)

}
