% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_month.R
\name{dif_month}
\alias{dif_month}
\title{Function for difference between dates}
\usage{
dif_month(date1, date2)
}
\arguments{
\item{date1}{object in Date format}

\item{date2}{object in Date format}
}
\description{
Function for difference between dates
}
\examples{

dif_month(as.Date('2022-12-15'),as.Date('2022-03-15'))

dif_month(as.Date('2022-12-15'),as.Date('2023-03-15'))

}
