% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile.R
\name{p10}
\alias{p10}
\alias{p25}
\alias{p75}
\alias{p90}
\alias{p95}
\alias{p99}
\title{Functions to calculate percentiles}
\usage{
p10(y, ...)

p25(y, ...)

p75(y, ...)

p90(y, ...)

p95(y, ...)

p99(y, ...)
}
\arguments{
\item{y}{Object with values}

\item{...}{as in quantile function}
}
\description{
Functions to calculate percentiles
}
\examples{

p25(1:3)
p75(1:4)
p99(0:100)

}
