% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_dependencies.R
\name{search_dependencies}
\alias{search_dependencies}
\title{Function for get dependencies recursively, excluding R base packages.}
\usage{
search_dependencies(packages, type = "Imports")
}
\arguments{
\item{packages}{vector with packages names}

\item{type}{type of dependencies. Like package_dependencies from tools package.}
}
\description{
Function for get dependencies recursively, excluding R base packages.
}
\examples{

search_dependencies(c('dplyr', 'shiny', 'data.table'))

}
