% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_log.R
\name{write_log}
\alias{write_log}
\title{Function for write messages in text file}
\usage{
write_log(message = "", file = "log.txt")
}
\arguments{
\item{message}{message to be saved in the log file

#' @examples
write_log('A Simple Test.')
readLines("log.txt")}

\item{file}{log file}
}
\description{
Function for write messages in text file
}
