% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfill.R
\name{zfill}
\alias{zfill}
\title{Fill leading zeros}
\usage{
zfill(x, len = max(nchar(x)))
}
\arguments{
\item{x}{Object with values}

\item{len}{Desired length}
}
\description{
Fill leading zeros until informed length. The default value is the largest value in x.
}
\examples{
zfill('a', 3) # returns '00a'
zfill(1:3, 3) # returns '001', '002', '003'
zfill(c(1, 12, 123)) # returns '001', '012', '123'

mtcars |>
  mutate(hp_fill = zfill(hp, 6))

# Mind cientific notation

options(scipen = 0)
1e6 |> zfill(largura = 9) # return "00001e+06"

options(scipen = 999)
1e6 |> zfill(largura = 9) # return "001000000"

}
