% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfill.R
\name{zfill}
\alias{zfill}
\title{Fill leading zeros}
\usage{
zfill(x, len = max(nchar(x)))
}
\arguments{
\item{x}{Object with values}

\item{len}{Desired length}
}
\description{
Fill leading zeros until informed length. The default value is the largest value in x.
}
\examples{
zfill('a', 3)
zfill(1:3, 3)
zfill(c(1, 12, 123))

mtcars$hp_fill <- zfill(mtcars$hp, 6)

# Mind cientific notation

options(scipen = 0)
1e6 |> zfill(len = 9) # return "00001e+06"

options(scipen = 999)
1e6 |> zfill(len = 9) # return "001000000"

}
